// src/types.ts
export interface AggregatedYearlyData {
    year: number;
    A: { type1: number; type2: number; type3: number; type4: number; total: number };
    B: { type1: number; type2: number; type3: number; type4: number; total: number };
    counts: { type1: number; type2: number; type3: number; type4: number; total: number };
}